requirements.txt
streamlit
pymongo
bson
requests
apscheduler
python-decouple
